requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
Flask-Session==0.5.0
marshmallow==3.20.1
Werkzeug==2.3.7
python-dotenv==1.0.1
PyMySQL==1.1.0  # Если используешь MySQL
